Added hand, card, card_factory (creates and returns a card), deck modules.
Added a player module with incomplete player logic, a table_controller module which should have methods for manipulating the table (also incomplete logic), and a game_controller, which should control the flow of the game by connecting the view and the model.

Some of these modules may belong in the Model package.